+++ examples/inputs/training_inputs_kestrel.yaml
@@ -8,7 +8,9 @@ dataset:
     # data_path: /projects/ssc/ahenry/wind_forecasting/awaken_data/filled_data_normalized.parquet
     # normalization_consts_path: /projects/ssc/ahenry/wind_forecasting/awaken_data/filled_data_normalization_consts.csv
     data_path: /projects/ssc/ahenry/wind_forecasting/awaken_data/SMARTEOLE_WakeSteering_SCADA_1minData_normalized.parquet
+    # data_path: /projects/ssc/ahenry/wind_forecasting/awaken_data/00_engie_scada_processed.parquet
     normalization_consts_path: /projects/ssc/ahenry/wind_forecasting/awaken_data/SMARTEOLE_WakeSteering_SCADA_1minData_normalization_consts.csv 
+    # normalization_consts_path: /projects/ssc/ahenry/wind_forecasting/awaken_data/SMARTEOLE_WakeSteering_SCADA_1minData_normalization_consts.csv 
     context_length: 9 # 120=10 minutes for 5 sec sample size
     prediction_length: 3 # 120=10 minutes for 5 sec sample size
     target_turbine_ids: # or leave blank to capture all
