+++ wind_forecasting/run_scripts/train_model_kestrel.sh
