+++ wind_forecasting/run_scripts/train_model_rc.sh
